requirements.txt
Flask
bcrypt
PyJWT